pyproject.toml
[project]
name = "playwright-python"
version = "0.1.0"
description = "this is playwright + pytest python testing project with allure and html reports"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-playwright = "^0.6.2"
allure-pytest = "^2.13.5"
requests = "^2.32.3"
playwright = "^1.49.1"
pytest-html = "^4.1.1"

